+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# `1.5`
+Fixed Julian day calculation
+
+Customizable (through command line):
+- Maghrib angle
+- Isha angle
+- Isha minutes
+- Elevation
+
+Added calculation methods:
+- MWL
+- ISNA
+- Egypt
+- Makkah (pass command line flag `--ramadan` if it is Ramadan to use properly)
+- Karachi
+- Tehran
+- Jafari
+
 # `1.4`
 Fixed Asr time calculation